Switch aggregation of JobRequest.started_at to an explicit call
We had a fairly narrow use of JobRequests previously, with every use
exposing the aggregated started_at from its related Jobs.  We now want
to add JobRequests to the staff area where not every use of them will
need that value and we don't want to pay the cost of computing it
implicitly anymore.