Add searchable filters to the JobRequestList page
To use select2, or any dropbox element, we need to wrap the entire page
in a form so we're also capturing the search box in the right-hand
column.  This allows us to keep the ability increase/reduce filters
without them overriding each other, and to search within any given set
of filters.