Add a command to backfill Job.run_command from existing JobRequests
This process takes ~30 minutes locally, so is likely to take much longer
in production.  Therefore we don't want to hold up a deploy waiting for
it to run in a migration.

I've added a progress bar so we can get an expected runtime and monitor
it as it goes.

I expect this to be removed once the backfilling is done.