Fix JobRequest.objects.with_started_at
with_started_at was previously overriding the default QuerySet for
JobRequests, hence super().get_queryset() in the method.  However, with
this left in it had the unexpected side-effect of removing any previous
filtering.

So a QuerySet in the form

    some_workspace.job_requests.with_start_at()

ended up with all JobRequests instead of just those for some_workspace.