Allow JobRequest's completed_at and runtime to query during render
Both of these methods are tricky to pull out of template rendering
because of complexities to their implementation.

completed_at relies on both getting the most recent job and checking the
aggregated status for the JobRequest before returning the last job's
completed_at value.

runtime is used on both detail and list pages and while it could be
pulled out into the view for detail pages, it gets a lot trickier for
list pages.  I tried a couple of routes and they both ended in complex
custom objects/dicts, so I've opted to mark these as allowed as the
pragmatic choice for ease of use.