Pull JobRequest.is_invalid out into the only place it's used
This property was doing an extra query during template rendering, which
both wasn't clear from the property and since it was only used in one
place, the abstraction wasn't really doing us very much.  Since we're
now also forcing the jobs query to execute before rendering we can do
this query easily in the view.