Fix the collect-me-maybe script regression.
This had been previously fixed, but was broken again in
6832a34e41b3ac5db021bd25a56d398bbc67f3b6, for 2+ months.

The collect-me-maybe.sh script relies on strings being output by
`manage.py print_settings`. If they are path objects, it outputs the
repr for them, which is not valid path.

To prevent this regressing again, I have added comment and a validation
check to the script.