Switch will_notify back to the custom form of form_radio component
will_notify is a BooleanField under the hood, instead of a
choices-supporting Field, so it doesn't provide the usual .choices()
method from which to get some values from.  In moving this use of the
component to use the BoundField instance I unwittingly used the static
field.value (always either True/False depending on context the page was
loaded in).

This reverts that change, explicitly using the strings True/False.