Run docker image as non-root user
This user switch built in to the image, and thus decided at *build
time*. This means we can't accidentally misconfigure dokku and run it as
root again.

In dev, it uses the developers local uid:gid. Because we run
collectstatic in the dev entrypoint, that meant the dev volume set up
needed a slight tweak. Rather than a docker volume, it's now a bind
mount volume. This just ensures that the user has the correct
permissions to write collectstatic output there. It also means we can
easily poke around in the docker-built assets if needed.

In prod, this is an arbitrary uid:gid, currently 10001.
This uid:gid should be set up in the production environment and given
ownership/perimssions to appropriate directories. This is not for
collectstatic (which is built into the image) but instead for releases
and other files that may get written to disk.

Part of https://github.com/opensafely-core/security/issues/15