Add user's roles page to the staff area
This adds a Roles page for each user to the staff area, listing global,
org, and project roles assigned to a user.

Changing a user's global roles has now moved to this page.

The edit and remove buttons for projects will redirect a user back to
the roles page on form submission.

While we can technically assign roles to an OrgMembership we don't
currently do or use that across the site.