Remove the maybe-this-is-a-workspace code from OrgDetail
This also allows us to revert back to it being a fairly simple
DetailView.  As per the comment this existed because we had no redirects
framework when we moved workspaces from the top level of our URLs.  This
happened over 2y ago and this view is no longer at that point in the URL
hierarchy either, so the code almost certainly hasn't been exercised in
some time.