Remove unused roles
When we designed our roles&permissions system we also designed the roles
(and permissions) around what we were currently doing and the system we
wanted to have when looking forward.  We haven't ended up using the
roles removed in this commit so far, and it's been over 2 years.  There
are no users using them in production either so they're simple to
remove.

Should we change our mind later they're simple to re-add.