Remove new release flow feature flag
This removes the 'new release flow' feature flag and the list of feature flags
from the project detail page in the staff area, since it was the only feature
flag in use.

The field has been removed from the Workspace edit form, which means any
changes to workspaces will not alter the uses_new_release_flow field.