Add a retry to GitHubAPI.delete_repo
We only use this for verification tests, and they can run fast enough
that GitHub hasn't finished creating the repo we're asking them to
delete.  The hope is a 1 second back off will fix this.  They don't give
us any backoff window for this operation so we're having to pick the
backoff ourselves.