Avoid extra queries executed by ModelChoiceField
This is a bit more work but avoids a query somewhere deep inside the
mechanics of ModelChoiceField.  I couldn't pinpoint the query, a count()
call as part of invoking `choices`.  ModelChoiceIterator defines a
__len__ which does a count, but it's not clear what's actually invoking
that when choices is called, particularly as choices implementation is
spread across the parent class, ChoiceField, and ModelChoiceField.