Fix applications test
We're using zen_queries' fetch() in the view to force evaluation of the
QuerySet.  Under the hood this calls QuerySet._fetch_all() to populate
its result cache and prefetch any necessary objects.  However, when
pagination is turned on the Paginator also adds an SQL OFFSET to the
query to get the relevant objects for the current page, which appears to
trigger the conversion to a list.