Set `database_name` based on project permissions
We've taken the deliberate decision here to use a hardcoded list of
project numbers committed to the codebase, rather than storing a new
set of permissions in the database. In the short-term we don't
anticipate needing to add to this list. We can revist the implementation
as and when that proves necessary.

As well as simplicity this approach has the advantage of creating a
natural public audit trail in the form of the commit history, where
changes to the database would be opaque.