Use Q objects to build up a [does not] have roles query
This lets us have one construction that we either filter or don't.
Django's .filter() and .exclude() have different semantics for how they
produce AND/OR queries.

.filter() uses AND regardless of multiple kwargs or chaining, while
.exclude() produces OR statements with multiple kwargs and AND
statements when chaining, making using these methods in combination
confusing to the reader.

Since we know we want the same query as either a filter or an exclude
this change builds up the query with Q objects (so we can explicitly
control the AND/OR statements generated) and then calls the appropriate
QuerySet method based on the input value.

This also removes the __exact lookup because it doesn't change the
resulting SQL.  It's unclear if this is because we're querying an array
field or because Postgres has optimised away the need for using IS.