Check for org and project roles by annotating the sum of them to a user
When filter (or excluding) users based on "has org/project roles" the
SQL query was incorrectly picking up users because they had _any_
membership with no roles, not all memberships with no roles.

By getting a sum of all role types annotated onto the user we can check
across all membership types for a given user.

This also adds a use case to the tests to capture this type of user.