Refactor running tests with coverage.
Added scripts/test-coverage.sh to handle the set up and teardown of
running test with coverage more robustly.

This allows us to keep the same test suite invocation across in local
and docker. Previously we'd had to duplicate the command pass it in as
single bash statement. This was causing some problems.

Also allowed us to simplify justfile and remove the need for test-dev,
and should hopefully fix xdist in CI.