Simplify interactive integration test
Now that the creation of the git commit has moved into interactive_templates,
we no longer need to test the commit creation here. The interactive_templates
tests are also checking run_all. We just need to confirm that we're
constructing everything correctly and that we're triggering the commit to be
created with the right information.

This should take around 40s off the test time and remove a fairly significant
dependency.