Only send report notifications if we create one successfully
Should creation of a report fail we want to avoid sending out
notifications about the analysis request.

create_report() is also wrapped in a transaction since it touches more
than one db object.  This change leaves that transaction alone so any
caller of it does not need to worry about it happening correctly, and
Django will handle the stacked transactions for us.