Set the charset to utf-8 when serving released files directly
When serving HTML files containing smart-quotes they were being mangled
when rendered.  Both reports and the outputs-viewer use this code path.
outputs-viewer did not mangle the quotes, but reports did, and when
viewing the response directly in a browser the quotes were still
mangled.

While response.charset was already set to utf-8, I also needed to add it
explicitly to the Content-Type header for the browser to fix the smart
quote rendering.

It's possible outputs-viewer was setting the charset to utf-8 in some
way I could not find.