Ingest per-job aggregate metrics
Job runner is now returning summary statistics for jobs (see
[PR](https://github.com/opensafely-core/job-runner/pull/686)). This allows us
to see things like the average and peak CPU and memory. We expect these fields
to change and be added to as we learn what's useful (or not), so the field has
been created as a JSON field to give us plenty of flexibility.