Push creation of ProjectMembership objects through the add() command
We want all users to create ProjectMembership objects with
project_members.add() so we raise an exception to point them in that
direction.  The override kwarg exists as an escape hatch in case someone
still needs to use it directly.

ProjectMembershipFactory has been removed in favour of the
project_member, and project_members, fixtures.  Again to make it clear
what a developer should be reaching for.