Display per-job metrics on Job Detail page
This adds a "Job metrics" card to the Job Detail page for users with one
or more roles. This card displays CPU and memory usage statistics, when
available (see below).

In addition to mean and peak, which are displayed here, Job Runner
collects sample and cumsum for both CPU and memory usage statistics
(opensafely-core/job-runner#686). However, I don't think sample and
cumsum usage statistics are useful to users: sample, because we (and so
they) don't know when the sample was taken; cumsum, because we (and so
they) don't know how many samples were taken (and if we did, then all we
would be able to compute would be the mean, which Job Runner has
computed for us).

As #3998 states, usage statistics are not available for historic jobs
(i.e. prior to 17/01/2024). In this case, `job.metrics == None`. They
are also not available for some other types of job. In these cases,
`job.metrics == {}`. We don't distinguish these cases, displaying "-" to
users in both cases.

Closes #3998