Refactor token generation and validation logic
After discussion with George, moving the logic from living in the model,
to in its own module. This moves in the direction of domain
organisation, and also links together the generation and validation with
the sending of related emails, as well as the logic for looking up the
user via email or username.

The next commit will add an additional callsite for validating tokens,
and this allows the logic (validate and email) to live in one place.