Add token authentication API for use by airlock
Adds it in the releases API, as thats what it is used for. We may want
to move it later.

The tricky part is building a list of workspaces the given user has the
correct permissions in. This is hard/impossible to do in the db, so we
have to do it in the app.  This is probably ok, as it will be a low
traffic API, and is already not available on the public IP.