Add create_backend command
It prints the auth token. Motivating use case was to set up a test
backend from cli for local tests. Allows you to add an existing user to
this backend, for convenience.

Replaced the janky script in the local release-hatch setup.

We want to use to automate setting up a local job-server in airlock.