Make Workspace.repo required
This is a leftover from moving from only storing the repo's URL to using
a model.  All workspaces in production currently have a repo, and this
ensures we flag the need for a repo to anyone creating workspaces in an
unexpected way.  It also ensures the various codepaths which assume
Workspace.repo exists aren't built on a false assumption.