Add create_user management command
This adds a new `create_user` management command, to aid in easily
creating users/roles in local development and testing. Motivated out of
consuming job-server as a service dependency in airlock.

This did require a change to way in which users are created in
auth_pipeline.py, which always attempted to create users when they log
in. Instead, when signing in for the first time with github auth, we
check if the username exists first. This allows a user that
already exists in the db, but doesn't have github linked, to be linked
on first login with Github.

This is very useful for setting up test intances of job-server, as it
allows us to script the creation of users ahead of actually logging in.

Users in django.contrib.auth are created with special
`create_user` function, which means we cannot user the
standard `get_or_create()` db model function. So we added an
implementation of `get_or_create_user()` that provides the same
experience. In addtion, we allow conditional updating of existing users,
as we want this in the github auth, but may only optionally want it on
a per field basis in the `create_user` command.