Don't update sub-dependencies
The `allow` [setting](https://docs.github.com/en/code-security/dependabot/dependabot-version-updates/configuration-options-for-the-dependabot.yml-file#allow) lets us configure what dependencies are updated (applies to both version and security updates). This change moves us back to the default of updating explicitly defined dependencies only. This will reduce the noise we're getting from dependabot and reduces the chance of an upstream dependency being updated to a version that's incompatible with one of our dependencies.