Fix created/deleted
Previously, all applications were labelled "Submitted" in the
applications list. However, in each case the datetime recorded when the
application was created.

Here, we change "Submitted" to either "Created" or "Deleted". In the
former case, we (still) use the creation datetime. In the latter case,
we use the deletion datetime.