Add `org_members.update_roles` command
Although we don't audit `OrgMembership.roles`, we still need a place to
update this field. The place should be separate from the `User` model,
because if we audit this field, then we'll need to distinguish between
the subject and object user. A new command meets this requirement, and
is consistent with `project_members.update_roles`.