Custom create_superuser
django uses `User.USERNAME_FIELD` to derive what fields are required for
the `mangage.py createsuperuser` but not for the built-in
`UserManager.create_superuser()` which expects a `username` regardless
so if this is changed to not include `username` then one will encounter
a missing arguent error.

Since we set the username field to `email` to support OSI user creation
we can/must set the username to email for superusers created
with the createsuperuser management command.

This also allows us to set the staff and superuser boolean flags
(which may soon be retired) and also assign the `CoreDeveloper` role
to ensure an adequately permissioned "staff" user is created by this
command.