Show lead org as default for projects
`Project.org` was replaced by `Project.orgs` when we moved from a
one-to-many to a many-to-many relationship between orgs and projects.
However, we need `Project.org` in several locations. For convenience, it
makes sense to show the lead org in these locations.

We expose `Project.org` as a property, meaning that an extra query is
executed each time a project's lead org is requested. Hooking properties
into [`prefetch_related`][1], which would reduce the number of extra
queries, is non-trivial. Should we bother?

The answer depends on how many extra queries we expect. Let's take the
`yours.ProjectList` view as an example
(https://jobs.opensafely.org/projects/). At present, the maximum number
of projects per user is 20 - held jointly by @alexwalkerepi and
@LFISHER7. For these users, this view executes 26 queries: five for page
setup, one for the list of projects, and 20 for the lead org for each
project. Locally, these queries execute in roughly 40ms. If we don't
expose the property, then this view executes six queries in roughly
10ms. So, the worst case for this view is a four-fold increase in
execution time.

20 projects per user is the maximum, though. At present, the median is
two projects per user (for users that are associated with at least one
project). So, hooking the property into `prefetch_related` wouldn't
change much for many users. We shouldn't bother!

[1]: https://docs.djangoproject.com/en/5.0/ref/models/querysets/#prefetch-related

Fixes #4136