Test atomicity of project-members commands
The commands in `jobserver.commands.project_members` are decorated with
`atomic` to ensure that writing to the DB results in either complete
success or complete failure. Generally, we'd rely on Django to test
`atomic`. However, because these commands are also adding entries to the
audit log, and because the semantics of the audit log are especially
important to us, we also test their atomicity.

We use `monkeypatch.context()` to be very specific about when we raise
`IntegrityError`s. We don't, for example, want to interfere with the
factories. Whilst we could locate `raise_integrity_error` in a utils
module, we don't because we still need to import `IntegrityError`
wherever the function is called.

Closes #4127