Assert on num db queries for views and templates
Our index page, whether logged in or not, makes quite a few calls to the
database, yet we want it to remain fairly performant. This adds some assertions
into the existing unit tests that check the view. It also adds some tests using
the Django Client to render the template and assert on the number of database
queries we're making using both the view and the template. That way, if we
accidentally change this behaviour by updating the template, we'll know. We may
want to expand these tests over time to ensure we're rendering the template
correctly or to check the number of db queries in different scenarios.