Refactor modal component to inline the js
Previously, using a modal component required two separate parts:

 1) a button component with data-modal=id attribute
 2) a modal component with the dialog contents

It also required a global include of modal.js, which would hook up the
open/close buttons.

This change makes the modal component a single artifact. It renders the
opening button by default. Similer to the card component, you can supply
the text/variant for the default button, or use your own custom_button
fragment.

It also inlines the script for setting up the open/close events. This
has two main benefits:

a) you don't need to remember to include the separate modal.js
b) it works with htmx partial loading, as the event wiring are included
in the content.

I have updated the documentation, and the one place we use the modal
currently, and manually tested both.