Add ImmutableModelMixin to ProjectMembership
This makes ProjectMembership an immutable model: it's no longer possible
to save or delete instances of this model without either raising an
error or passing override=True (the choice of keyword is consistent with
ProjectMembershipManager). ProjectMembership still has a part-mutable
query set, which we will fix in follow up commits.