Allow `OLD_SECRET_KEY` for secret key rotation
When rotating the secret key, all users would get logged out at once.

Persisting the old key for a while gives users chance to access the
site, and then get switched to sessions encrypted with the new key.

In cases where we are performing a scheduled secret key rotation — as opposed
to responding to a compromised secret key — this minimises the
disruption to users.

The workflow here is:

* Add the current secret key as an environment variable called
  `OLD_SECRET_KEY`.
* Change the `SECRET_KEY` to a new value, and ensure the server is
  restarted.
* After some time has elapsed, delete the `OLD_SECRET_KEY` environment variable,
  and ensure the server is restarted.