Remove double-click prevention script
The double-click prevention was disabling any submit button for 30
seconds. This isn't needed for the majority of forms, and causes
problems if a form has validation errors (because the button is disabled
for so long it seems like the form can't be submitted even after errors
are corrected).

Since we're not sure it's needed, removing it altogether for now.