Add "Short data report" study type
Adding a new study type means making several changes. However, these
changes are not especially meaningful on their own; hence making them in
one commit.

Of least concern should be the changes to the rubrics: these are copied
from #3874.

The changes to the models, which encompass the migration, the factory,
and the mutability test, ensure that an application for a "Short data
report" is persisted and that such an application is identified
correctly. The changes to `form_specs`, which encompass the integration
test, ensure that the application flow is correct. That the flow is
correct may not be evident from the integration test. I hope the
comments help.

Closes #3874