Make purpose of expression explicit
The purpose of the expression is to indicate whether a user is a member
of a project. We make the purpose explicit by assigning the result of
calling the expression to a variable. We also rename the associated
test, which isn't about permissions (in the "roles and permissions"
sense) at all.