Add the `role_factory` fixture
This is a fixture for dynamically creating a role with a given set of
permissions. We will use this fixture to create single-permission roles
in tests, replacing the existing multi-permission roles that are also
used in production. As well as decoupling our production roles from our
tests, using this fixture will ensure our tests are more targeted.