`ProjectDeveloper` -> single-permission role
We replace `ProjectDeveloper`, a multi-permission role used in
production, with a dynamically-created, single-permission role. Doing so
makes these tests more targeted, as we can be confident that
`ProjectEdit` is restricting users based on the `project_manage`
permission, and not on one of the other permissions that are also
associated with `ProjectDeveloper`.