Return existing role, if possible
Previously, calling `role_factory` would instantiate a role, even if a
role with the same name (and hence the same permission) existed. Now, it
returns an existing role, if possible. Whilst doing so improves
performance slightly, the main benefit is that doing so makes it clear
that `role_factory` doesn't tear-down between tests.