Add github_api methods to close issue
Changes state of an issue to closed, with an optional comment.

This also adds options to return the most/least recently created
matching issue (defaults to latest) and to fetch only open/closed
issues (default to all).

Tests using the real github api use the existing issues in the testing
repo to avoid created lots of new issues every time the tests run
(especially as issues can only be closed, not deleted). There's already
a test that creates an issue, and there are >5K issues in the repo, so
closing the oldest one in a test shouldn't cause any problems.