Handle unknown airlock event types
Airlock posts to the events endpoint for a change in status to
REQUEST_APPROVED. Currently we don't send any notifications or
action any github issues for this status, but we weren't handling
it; now it accepts a REQUEST_APPROVED status, but doesn't action
any notifications; for now we'll keep this no-action path as it's
inexpensive and might be required later.

We also now catch an unknown event type received from Airlock and
return an error response.