Switch `html_utils.clean_html`'s cleaner
We switch from `lxml.html.clean.Cleaner` to `nh3.clean`. Doing so will
allow us to remove the former (in a follow-up commit).

Whilst `nh3.clean` says that it conforms to the HTML5 specification, it
adds unnecessary `tbody` elements to `table` elements, which doesn't
conform.[^1] However, we accept that it adds unnecessary `tbody`
elements, modify the tests, and move on.

Closes #4254

[^1]: https://html.spec.whatwg.org/multipage/tables.html#the-tbody-element