Include project status in user details endpoint
This sends a boolean "ongoing" status rather than the specific
Project status. This data is expected to be consumed by Airlock,
which just needs to know whether a project is ongoing or not in
order to determine whether release requests can be created. In
case the specific statuses change in job-server (and there's some
comments in the model that suggest that they might in future), we
don't want Airlock to be looking for exact job-server controlled
strings.