Test application (staff) detail num db queries
This counts the maximum number of db queries performed in the view and in the
template.  We don't expect that number to increase, but it might if we
accidentally make a change to the template. Now we will have a failing test to
tell us that's happened (and what queries are being made).

Note: the number of db queries performed in the view varies between test runs
and is either 18 or 19. This was the case previously, so nothing has been
changed in this commit.