Replace zen queries with Django TemplateResponse
We can be confident that this isn't producing any extra db queries in the
frontend because we now have a unit test for that.

We've replaced the fetch(queryset..) with a call to list() as this forces the
queryset to be
[evaluated](https://docs.djangoproject.com/en/5.0/ref/models/querysets/#when-querysets-are-evaluated)
before sending it to the template, so we aren't triggering any extra db calls
in the template.