Modify workflow to run from within job-server repo
Instead of having interactive-templates create a PR when a new release
is available, check from within this repository on a regular basis.

Having this be triggered as a "pull" notification rather than a "push"
notification is not quite as refined in terms of getting updates.

However, it removes the use of a personal access token. Generally, we're
finding those unwieldy to manage as an organisation: see ebmdatalab/metrics#194.

See opensafely-core/interactive-templates#118 and
opensafely-core/interactive-templates#377 (where this workflow was
almost working in interactive-templates, except for requiring a new
personal access token).