Fix `ApplicationList` view
4371e7e modified `ApplicationList.get_queryset` to return a list, rather
than an evaluated query set (the result of passing an unevaluated query
set to django-zen-queries' `fetch` function). Doing so meant that
`ApplicationList.context_object_name` was not set, which meant that the
corresponding template no longer had `application_list` in its context,
which meant that the list of applications was empty.

Thanks, @LiamHart-hub for reporting this issue.