Return queryset from `.get_queryset`
Previously, `ApplicationList.get_queryset` returned a list, which is a
Django idiom for evaluating (lazy) querysets ([1][]). However,
evaluating a queryset in this way means
`ApplicationList.get_context_object_name` returns `None`, rather than a
suitable name for the context object, such as `"application_list"`. This
is undesirable, because the view's associated template depends on
`application_list`.

Returning a queryset increases the number of queries made when the view
is rendered from one to two. This number isn't related to the number of
applications in the applications list (there are actually five). One
query returns the applications and one query relates to Python Social
Auth. An alternative would be to mimic django-zen-queries' `fetch`
function by calling `qs._fetch_all()` ([2][]). However, I'm feeling very
un-Zen about "internal use" methods and don't think the extra query is
worth the bother.

Another alternative would be to set
`ApplicationList.context_object_name` to `"application_list"`, which was
what we did in 8572f35. Doing so feels like we're adding yet more
accidental complexity: a method called `.get_queryset` that doesn't get
a queryset and a property called `context_object_name` to make up for
it. So, let's not do that either.

[1]: https://docs.djangoproject.com/en/5.0/ref/models/querysets/#when-querysets-are-evaluated
[2]: https://github.com/dabapps/django-zen-queries/blob/2.1.0/zen_queries/utils.py