Make django-debug-toolbar a development dependency
Although making django-debug-toolbar (DDT) a development dependency
introduces branches in `settings.py` and `urls.py`, which we have tried
to avoid in the past (e.g. bfdc66f), it nevertheless makes our
production deployment more secure. Depending on DDT only in development
is also clearer than depending on it in production and in development,
but indirectly disabling it in production via `settings.INTERNAL_IPS`.

This commit takes inspiration from opensafely-core/airlock#131 and
isolates specific DDT-related settings from general debug-related
settings. It does this with the `DJANGO_DEBUG_TOOLBAR` environment
variable and the `DEBUG_TOOLBAR` setting (the `DJANGO_` prefix is
dropped from the setting, as it's unnecessary). For safety, the setting
is `False` by default. Because seeing DDT is a good reminder that the
browser tab you're looking at is pointing to development and not
production, DDT is enabled in development by hard-coding the environment
variable in `just run`.

Closes #4441