Test number of db queries on login
This counts the number of db queries performed in the view and in the template
for both a GET and POST request on login. We don't expect that number to
change, but it might if we accidentally make a change to the template. Now we
will have a failing test to tell us that's happened (and what queries are being
made).