Add cron checkin events to sentry parsing test
These json files were dumped from the debugger whilst running a function
that purposefully fails which has been decorated with `monitor`.

The first (event 3) is a "in progress" checkin which is sent as the
function is entered and the second (event 4) is an error event.

The error event contains a float duration field which had previously
caused the `parse()` method to fail.