Move health check to `app.json`
Fixes #4489.

`WAIT` was originally the time to wait before running the first check as
explained, for example, in this old Dokku documentation page:

https://dokku.com/docs~v0.3.26/checks-examples/

This now corresponds to `initialDelay` in the `app.json` health check:

> Number of seconds to wait after a container has started before triggering the healthcheck.

(`wait` in `app.json` corresponds to the time between check attempts,
which doesn't, from my reading, seem to have been configurable in the
old `CHECKS` configuration. That is, a subsequent check would occur
immediately after a failed check.)