Inline the `first` dependency
This is an intermediate step to eventually removing this code.

By including the code, we can:

* remove the external dependency installation
* record the existing behaviour without having to refer to, or even rely
  on the availability of the upstream code

`first` is a function in a package also called `first`. Briefly, as we
use `first()` in job-server, it returns the first truthy value from an
iterable.  `first` can also take an optional `default` argument to
provide if no such value is found. If no default is specified, the
default is `None`.