Replace zen queries with Django TemplateResponse
This view was broken in commit a8590ba when a zen queries exception was
removed from the project.org property. So, this change adds a test and removes
zen queries from the view and template to fix the problem. Calling
`response.render()` in the test was enough to demonstrate that it was broken.