Placate coverage for tests
The coverage check complains about branches not being executed. This is
because they never reach the end of the generator in each case; there is
always a value.

In these cases, this is possibly OK to ignore them, because we never
expect that we should get the alternate possible branch. There is
always a valid value found from the generator expression: otherwise
these tests would fail (they require a non-`None` value to continue).

See the coverage documentation:

https://github.com/nedbat/coveragepy/blob/2a7d9779add1e8fa3425a784347316f5d28a510d/doc/branch.rst#L120-L131