Temporarily disable the error warning filter
The error warning filter transforms warnings into errors. We're going to
make a change that will report more warnings, which we'll then fix. We
don't want the test suite to fail after we make the change, but before
we fix the warnings, so we temporarily disable the error warning filter