Add `ProjectFactory.org`
`Project.org` is a property, rather than a field, so we need to tell
factory_boy what to do with it when creating a new instance of
`Project`. To do so, `_project_collaboration_factory` needs to handle
the case where `extracted` is a single org, rather than a list of orgs.