Only show repo sign-off btn to permissioned users
bug: workspaces with private repos will show the "admin section" to all
users regardless of their permissions.

cause: the display logic is to check that the repo is private OR the
user has one of two relevant permissions.

fix: this changes the display logic to check only that the user has one
of two relevant permissions.
The template contains logic to only display the repo visibility change
button for private repos so it is not needed in the view.